Moved schema_tests to be part of unit_tests Moved some functions into application/deed/utils (they will be useful for the next task) Removed the now redundant schema_tests/utils Added unit test to run schema_tests
