--- schema_test.sh
@@ -1,1 +1,0 @@
-python3 schema_tests.py
